revisions crosshair implementation
