+++ Unity/Assets/Scenes/SampleScene.unity
@@ -156,7 +156,6 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   Reticle: {fileID: 1208500442}
   BulletPrefab: {fileID: 50889125771127516, guid: 8697fff0611cd3c4b8bd5548c22ffc64, type: 3}
-  _movementForce: 10
 --- !u!212 &428799269
 SpriteRenderer:
   m_ObjectHideFlags: 0
@@ -601,12 +600,12 @@ SpriteRenderer:
   m_SortingLayerID: 0
   m_SortingLayer: 0
   m_SortingOrder: 0
-  m_Sprite: {fileID: 21300000, guid: e2f632dd02b6d41958501172b35012b9, type: 3}
-  m_Color: {r: 1, g: 0.38597694, b: 0, a: 0.21568628}
+  m_Sprite: {fileID: 21300000, guid: 03ed46d7646564a35bb98a5ce30b5105, type: 3}
+  m_Color: {r: 1, g: 0.38597694, b: 0, a: 0.54509807}
   m_FlipX: 0
   m_FlipY: 0
   m_DrawMode: 0
-  m_Size: {x: 1, y: 1}
+  m_Size: {x: 1.28, y: 1.28}
   m_AdaptiveModeThreshold: 0.5
   m_SpriteTileMode: 0
   m_WasSpriteAssigned: 1
