+++ Unity/Assets/Sprites/crosshair063.png.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e2f632dd02b6d41958501172b35012b9
+guid: 03ed46d7646564a35bb98a5ce30b5105
 TextureImporter:
   internalIDToNameTable: []
   externalObjects: {}
